Separate ShardingRuleConfigurationSwapper and rename
